+++ core-ng-test/src/test/java/core/framework/test/TestModule.java
@@ -30,7 +30,7 @@ public class TestModule extends AbstractTestModule {
         load(new OverrideBeanTest());
 
         overrideBinding(HTTPClient.class, Mockito.mock(HTTPClient.class));  // in test context, override binding is defined before actual binding
-        bind(HTTPClient.class, HTTPClient.builder().tlsVersions("TLSv1.3", "TLSv1.2").maxRetries(3).retryWaitTime(Duration.ofSeconds(1)).enableCookie().build());
+        bind(HTTPClient.class, HTTPClient.builder().maxRetries(3).retryWaitTime(Duration.ofSeconds(1)).enableCookie().build());
 
         configureDB();
         configureKafka();